Phase 116.1: Fix Clang 15 sign-conversion errors in span accessors
Fixed implicit sign conversion errors when using std::span with integer indices.
Changed loop variables from 'int' to 'size_t' where used as span indices.

Files modified:
- src/core/ldebug.cpp: Fixed 3 functions (luaG_getfuncline, nextline, changedline,
  collectvalidlines) to use size_t for span indexing
- src/objects/ltable.cpp: Fixed NodeArray::allocate overflow check
- src/serialization/lundump.cpp: Fixed loadConstants, loadUpvalues, loadDebug
  to use size_t loop variables

Critical fix in luaG_getfuncline: Changed loop from 'i < pc' to 'i <= pc' to
match original while(basepc++ < pc) semantics (process indices basepc+1 through pc inclusive).

All tests passing with GCC 13 and Clang 15 in Release and Debug modes.
Performance: 2.26-2.56s (well within target ≤4.33s)

+++ src/objects/ltable.cpp
@@ -111,7 +111,7 @@ public:
       // Large table: allocate Limbox + Node[]
       // LAYOUT: [Limbox header][Node array of size n]
       // Verify no overflow in size calculation
-      if (n > (MAX_SIZET - sizeof(Limbox)) / sizeof(Node)) {
+      if (static_cast<size_t>(n) > (MAX_SIZET - sizeof(Limbox)) / sizeof(Node)) {
         luaG_runerror(L, "table size overflow");
       }
       size_t total = sizeof(Limbox) + n * sizeof(Node);
